Moved Navigator and reduced dependency in StockChart.
