+++ ts/Core/Chart/StockChart.ts
@@ -29,16 +29,13 @@ import type { SeriesTypePlotOptions } from '../Series/SeriesType';
 import type SVGElement from '../Renderer/SVG/SVGElement';
 import type SVGPath from '../Renderer/SVG/SVGPath';
 
-import A from '../Animation/AnimationUtilities.js';
-const {
-    animObject
-} = A;
 import Axis from '../Axis/Axis.js';
 import Chart from '../Chart/Chart.js';
 import F from '../../Core/FormatUtilities.js';
 const { format } = F;
-import D from '../DefaultOptions.js';
-const { getOptions } = D;
+import DO from '../DefaultOptions.js';
+const { getOptions } = DO;
+import NavigatorDefaults from '../Navigator/NavigatorDefaults.js';
 import { Palette } from '../../Core/Color/Palettes.js';
 import Point from '../Series/Point.js';
 import Series from '../Series/Series.js';
@@ -58,9 +55,6 @@ const {
 } = U;
 
 import '../Pointer.js';
-// Has a dependency on Navigator due to the use of
-// defaultOptions.navigator
-import '../Navigator.js';
 // Has a dependency on Scrollbar due to the use of
 // defaultOptions.scrollbar
 import '../Scrollbar.js';
@@ -149,7 +143,7 @@ class StockChart extends Chart {
             // navigator is enabled (#1090)
             navigatorEnabled = pick(
                 userOptions.navigator && userOptions.navigator.enabled,
-                (defaultOptions.navigator as any).enabled,
+                NavigatorDefaults.enabled,
                 true
             );
 
@@ -378,14 +372,13 @@ function getForcedAxisOptions(
     chartOptions: Partial<Options>
 ): DeepPartial<AxisOptions> {
     if (type === 'xAxis') {
-        const defaultOptions = getOptions(),
-            // Always disable startOnTick:true on the main axis when the
-            // navigator is enabled (#1090)
-            navigatorEnabled = pick(
-                chartOptions.navigator && chartOptions.navigator.enabled,
-                (defaultOptions.navigator as any).enabled,
-                true
-            );
+        // Always disable startOnTick:true on the main axis when the navigator
+        // is enabled (#1090)
+        const navigatorEnabled = pick(
+            chartOptions.navigator && chartOptions.navigator.enabled,
+            NavigatorDefaults.enabled,
+            true
+        );
 
         const axisOptions: DeepPartial<AxisOptions> = {
             type: 'datetime',
@@ -899,7 +892,7 @@ addEvent(Chart, 'update', function (
     // case (#6615)
     if ('scrollbar' in options && this.navigator) {
         merge(true, this.options.scrollbar, options.scrollbar);
-        (this.navigator.update as any)({}, false);
+        this.navigator.update({});
         delete options.scrollbar;
     }
 });
